libcamera: Fix image cropping
Image dimesnsions and their crop dimensions need to be multiples
of 16. I don't believe that the VM670 supports any image sizes
where this isn't the case (CAF uses 5MP and SVGA as examples).
So, for now, use CEILING16 to ensure compliance in the crop
path only.

Also obliterate the jpegPadding variable since it's not being used.

Closes inferiorhumanorgans/android_device_lge_thunderc#17.

Change-Id: Ib442e62b151900f68177a594cf1536b6eba62f45