libcamera: Fix double tap in camera app leads to FC.
The FC only happened if you were to double tap before fiddling with
the zoom slider.  If you set the zoom manually beforehand, double tap
to zoom min/max worked just fine.

Problem was that the max zoom step was initially being set to the
camera's max zoom level which works well if you're using a ZOOM_STEP
of 1.  This is a bit silly since it's still utterly dependent upon
the native max zoom not changing beyond ZOOM_STEP*MAX_ZOOM_LEVEL.
It's probably worth rethinking setting/overriding this in the big
set prefs code block in the first place.

Change-Id: I0131878cc39eb04a105add4341bba6c62b6e2b0c