libaudiopolicy: Choose only one output device w/ STRATEGY_SONIFICATION
STRATEGY_SONIFICATION signifies that we should force output through
the main speaker.  STRATEGY_MEDIA_SONIFICATION means we should go
through the hierarchy and find the right device.  Thing is the
MSM7k audio bits don't seem to support multiple output devices.
Soo... let's specify *ONLY* the speaker (and not the speaker plus
whatever else is plugged in) when we get to the STRATEGY_SONIFICATION
case.

Closes inferiorhumanorgans/android_device_lge_thunderc#45.

Change-Id: Ifa0503a19e563e31fd4d382771d52ddea14c8429